requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
requests==2.32.5
SQLAlchemy==2.0.43
PyPDF2==3.0.1
python-docx==1.2.0

